requirements.txt
Flask==3.1.2
DrissionPage==4.1.1.2
requests==2.32.5
tqdm==4.67.1
cozepy==0.20.0
pandas==2.3.0
openpyxl==3.1.5
scikit-learn==1.5.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-CORS==4.0.0
PyJWT==2.8.0
Flask-RESTX==1.3.0